Index thumbnail links
Added style to make cursor, used jQuery to have whole thumbnail
clickable. It's not perfect since it doesn't appear as a link to
the browser, but it gets the job done and is touch friendly
without breaking accessibility (as would be the case for putting
one big <a> around the div with the image).